requirements.txt
python-dotenv==1.0.1
pydantic==2.7.1
SQLAlchemy==2.0.30
